requirements.txt
discord.py==2.2.2
mcrcon==0.7.0
python-dotenv==1.0.0